refactor reindex job into a celery task
